@@ -5,11 +5,11 @@ import (
 	"log"
 	"os"
 
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 func main() {
-	var c interface{}
+	var c any
 	if err := json.NewDecoder(os.Stdin).Decode(&c); err != nil {
 		log.Fatal(err)
 	}
